Add more option to usePosition hook
