+++ src/portal/TooltipBody.tsx
@@ -8,8 +8,8 @@
 
 import React from 'react';
 
+import usePosition from '../hooks/usePosition';
 import ThemeContent from '../theme/ThemeContext';
-import calculatePosition from '../utils/calculatePosition';
 import Arrow from './Arrow';
 import Offset from './Offset';
 import Position from './Position';
@@ -26,19 +26,7 @@ const TooltipBody: React.FC<TooltipBodyProps> = ({ children, offset, position, t
     const contentRef = React.createRef<HTMLDivElement>();
     const anchorRef = React.createRef<HTMLDivElement>();
 
-    React.useLayoutEffect(() => {
-        const targetEle = targetRef.current;
-        const contentEle = contentRef.current;
-        const anchorEle = anchorRef.current;
-        if (!contentEle || !targetEle || !anchorEle) {
-            return;
-        }
-
-        const anchorRect = anchorEle.getBoundingClientRect();
-        const { top, left } = calculatePosition(contentEle, targetEle, position, offset);
-        contentEle.style.top = `${top - anchorRect.top}px`;
-        contentEle.style.left = `${left - anchorRect.left}px`;
-    }, []);
+    usePosition(contentRef, targetRef, anchorRef, position, offset);
 
     return (
         <>
