+++ src/portal/tooltipBody.less
@@ -12,11 +12,11 @@
     background: #000;
     border-radius: 4px;
     color: #FFF;
-    /*left: 0;*/
+    left: 0;
     max-width: 300px;
     position: absolute;
     text-align: center;
-    /*top: -9999px;*/
+    top: -9999px;
     z-index: 9999;
 
     &-arrow {
